Move consistent_trees_hlist frontend into consistent_trees.
