--- ytree/frontends/consistent_trees_hlist/__init__.py
@@ -1,17 +1,0 @@
-"""
-consistent_trees_hlist frontend
-
-
-
-"""
-
-#-----------------------------------------------------------------------------
-# Copyright (c) ytree development team. All rights reserved.
-#
-# Distributed under the terms of the Modified BSD License.
-#
-# The full license is in the file COPYING.txt, distributed with this software.
-#-----------------------------------------------------------------------------
-
-from ytree.frontends.consistent_trees_hlist.arbor import \
-    ConsistentTreesHlistArbor
